fix(runtime): share safe shell environment policy
