+++ crates/zeroclaw-runtime/src/tools/mod.rs
@@ -21,6 +21,7 @@ pub mod scoped;
 pub mod security_ops;
 pub mod send_message_to_peer;
 pub mod shell;
+pub(crate) mod shell_env;
 pub mod skill_http;
 pub mod skill_manage;
 pub mod skill_tool;
