+++ crates/zeroclaw-runtime/src/tools/skill_tool.rs
@@ -2,6 +2,7 @@
 
 use crate::platform::{NativeRuntime, RuntimeAdapter};
 use crate::security::SecurityPolicy;
+use crate::tools::shell_env::SAFE_SHELL_ENV_VARS;
 use async_trait::async_trait;
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -14,31 +15,6 @@ use zeroclaw_api::tool::{Tool, ToolOutput, ToolResult};
 const SKILL_SHELL_TIMEOUT_SECS: u64 = 60;
 /// Maximum output size in bytes (1 MB).
 const MAX_OUTPUT_BYTES: usize = 1_048_576;
-
-#[cfg(not(target_os = "windows"))]
-const SAFE_ENV_VARS: &[&str] = &[
-    "PATH", "HOME", "TERM", "LANG", "LC_ALL", "LC_CTYPE", "USER", "SHELL", "TMPDIR",
-];
-
-#[cfg(target_os = "windows")]
-const SAFE_ENV_VARS: &[&str] = &[
-    "PATH",
-    "PATHEXT",
-    "HOME",
-    "USERPROFILE",
-    "HOMEDRIVE",
-    "HOMEPATH",
-    "SYSTEMROOT",
-    "SYSTEMDRIVE",
-    "WINDIR",
-    "COMSPEC",
-    "PSModulePath",
-    "TEMP",
-    "TMP",
-    "TERM",
-    "LANG",
-    "USERNAME",
-];
 
 const MAX_TOOL_NAME_LEN: usize = 64;
 
@@ -238,7 +214,7 @@ impl Tool for SkillShellTool {
         cmd.env_clear();
 
         // Only pass safe environment variables
-        for var in SAFE_ENV_VARS {
+        for var in SAFE_SHELL_ENV_VARS {
             if let Ok(val) = std::env::var(var) {
                 cmd.env(var, val);
             }
@@ -641,12 +617,6 @@ mod tests {
         let result = tool.execute(serde_json::json!({})).await.unwrap();
         assert!(result.success);
         assert!(result.output.contains("hello-skill"));
-    }
-
-    #[cfg(windows)]
-    #[test]
-    fn skill_shell_safe_env_vars_preserves_powershell_module_discovery() {
-        assert!(SAFE_ENV_VARS.contains(&"PSModulePath"));
     }
 
     #[cfg(windows)]
